Widen `VarBinBuilder` offets during FSST compress (#7853)
## Summary

Fixes: #7833

I don't think that a more complicated solution here makes a lot of
sense. We narrow these offsets is the compressor anyways, so there is no
reason to do something more complicated than this.

Additionally adds the logic for comparing with arrow large binary arrays
that was missing before.

To be honest, I feel like it doesn't really make any sense to allow the
user to specify what the offsets and sizes types should be at all, I
think we can have the builder _always_ use the 64-bit types because we
know that they will be compressed by the compressor anyways.

## Testing

Adds some regression tests. I've ignored the big test because it
allocates too much memory.

You can run the ignored test with:

```
cargo t --release -p vortex-fsst varbin_compress_offsets_overflow_i32 -- --ignored --no-capture
```

+++ encodings/fsst/src/compress.rs
@@ -69,8 +69,12 @@ where
     I: Iterator<Item = Option<&'a [u8]>>,
 {
     let mut buffer = Vec::with_capacity(DEFAULT_BUFFER_LEN);
-    let mut builder = VarBinBuilder::<i32>::with_capacity(len);
+
+    // Offsets are widened to i64 because the cumulative compressed bytes can exceed i32::MAX for
+    // large inputs (see issue #7833). Per-string sizes still fit in i32.
+    let mut builder = VarBinBuilder::<i64>::with_capacity(len);
     let mut uncompressed_lengths: BufferMut<i32> = BufferMut::with_capacity(len);
+
     for string in iter {
         match string {
             None => {
